Initial commit, 0.1.0
Backend supports essential operations: the sell interface, an option to
top-up tags, authentication.

Adding and removing sellable items or tags is not yet implemented and
must be achieved by SQL, but these are one-time operations and as such
don't introduce a lot of friction, at least for the upcoming MVP test.